backend/pyproject.toml
[project]
name = "time-tracking-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.7",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pywebpush>=2.3.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
